+++ src/api/mod.rs
@@ -1,1 +1,3 @@
+pub mod font;
 pub mod syscall;
+pub mod vga;
