+++ src/api/vga/mod.rs
@@ -1,0 +1,5 @@
+pub mod color;
+pub mod palette;
+
+pub use color::Color;
+pub use palette::Palette;
