+++ src/api/vga/palette.rs
@@ -1,0 +1,3 @@
+pub struct Palette {
+    pub colors: [(u8, u8, u8, u8); 16]
+}
