+++ src/kernel/mod.rs
@@ -5,7 +5,6 @@ pub mod clock;
 pub mod cmos;
 pub mod console;
 pub mod cpu;
-pub mod fonts;
 pub mod fs;
 pub mod gdt;
 pub mod idt;
